JP-5401: Simplify the ComTaskExecution hierarchy: merge AdhocComTaskExecutionImpl into ManuallyScheduledComTaskExecutionImpl

+++ comu/com.energyict.mdc.device.data/src/main/java/com/energyict/mdc/device/data/impl/tasks/ManuallyScheduledComTaskExecutionImpl.java
@@ -5,13 +5,23 @@ import com.energyict.mdc.device.config.ProtocolDialectConfigurationProperties;
 import com.energyict.mdc.device.data.Device;
 import com.energyict.mdc.device.data.DeviceDataService;
 import com.energyict.mdc.device.data.exceptions.MessageSeeds;
+import com.energyict.mdc.device.data.tasks.ComTaskExecutionBuilder;
 import com.energyict.mdc.device.data.tasks.ManuallyScheduledComTaskExecution;
-import com.energyict.mdc.device.data.tasks.ManuallyScheduledComTaskExecutionBuilder;
 import com.energyict.mdc.device.data.tasks.ManuallyScheduledComTaskExecutionUpdater;
 import com.energyict.mdc.scheduling.NextExecutionSpecs;
 import com.energyict.mdc.scheduling.SchedulingService;
 import com.energyict.mdc.scheduling.TemporalExpression;
+import com.energyict.mdc.scheduling.model.ComSchedule;
+import com.energyict.mdc.tasks.BasicCheckTask;
+import com.energyict.mdc.tasks.ClockTask;
 import com.energyict.mdc.tasks.ComTask;
+import com.energyict.mdc.tasks.LoadProfilesTask;
+import com.energyict.mdc.tasks.LogBooksTask;
+import com.energyict.mdc.tasks.MessagesTask;
+import com.energyict.mdc.tasks.ProtocolTask;
+import com.energyict.mdc.tasks.RegistersTask;
+import com.energyict.mdc.tasks.StatusInformationTask;
+import com.energyict.mdc.tasks.TopologyTask;
 
 import com.elster.jupiter.domain.util.Save;
 import com.elster.jupiter.events.EventService;
@@ -24,6 +34,10 @@ import com.elster.jupiter.util.time.Clock;
 import com.google.common.base.Optional;
 
 import javax.inject.Inject;
+import java.util.Arrays;
+import java.util.Collections;
+import java.util.Date;
+import java.util.List;
 
 /**
  * Provides an implementation for the {@link ManuallyScheduledComTaskExecution} interface.
@@ -31,9 +45,13 @@ import javax.inject.Inject;
  * @author Rudi Vankeirsbilck (rudi)
  * @since 2014-06-30 (10:59)
  */
-public class ManuallyScheduledComTaskExecutionImpl extends SingleComTaskExecutionImpl implements ManuallyScheduledComTaskExecution {
+public class ManuallyScheduledComTaskExecutionImpl extends ComTaskExecutionImpl implements ManuallyScheduledComTaskExecution {
 
-    @IsPresent(groups = {Save.Create.class, Save.Update.class}, message = "{" + MessageSeeds.Keys.NEXTEXECUTIONSPEC_IS_REQUIRED + "}")
+    @IsPresent(groups = {Save.Create.class, Save.Update.class}, message = "{" + MessageSeeds.Keys.COMTASK_IS_REQUIRED + "}")
+    private Reference<ComTask> comTask = ValueReference.absent();
+    @IsPresent(groups = {Save.Create.class, Save.Update.class}, message = "{" + MessageSeeds.Keys.PROTOCOL_DIALECT_CONFIGURATION_PROPERTIES_ARE_REQUIRED + "}")
+    private Reference<ProtocolDialectConfigurationProperties> protocolDialectConfigurationProperties = ValueReference.absent();
+    @IsPresent(groups = {SaveScheduled.class}, message = "{" + MessageSeeds.Keys.NEXTEXECUTIONSPEC_IS_REQUIRED + "}")
     private Reference<NextExecutionSpecs> nextExecutionSpecs = ValueReference.absent();
 
     @Inject
@@ -42,33 +60,60 @@ public class ManuallyScheduledComTaskExecutionImpl extends SingleComTaskExecutio
     }
 
     public ManuallyScheduledComTaskExecutionImpl initialize(Device device, ComTaskEnablement comTaskEnablement, ProtocolDialectConfigurationProperties protocolDialectConfigurationProperties, TemporalExpression temporalExpression) {
-        this.initializeFrom(device, comTaskEnablement);
+        super.initializeFrom(device, comTaskEnablement);
+        this.setComTask(comTaskEnablement.getComTask());
+        this.setProtocolDialectConfigurationProperties(comTaskEnablement.getProtocolDialectConfigurationProperties().orNull());
         this.setProtocolDialectConfigurationProperties(protocolDialectConfigurationProperties);
-        this.setNextExecutionSpecsFrom(temporalExpression);
+        if (temporalExpression != null) {
+            this.setNextExecutionSpecsFrom(temporalExpression);
+        }
         return this;
+    }
+
+    public ManuallyScheduledComTaskExecutionImpl initializeAdhoc(Device device, ComTaskEnablement comTaskEnablement, ProtocolDialectConfigurationProperties protocolDialectConfigurationProperties) {
+        return this.initialize(device, comTaskEnablement, protocolDialectConfigurationProperties, null);
     }
 
     @Override
     protected void postNew() {
         this.recalculateNextAndPlannedExecutionTimestamp();
-        super.postNew();
+        if (this.isAdHoc()) {
+            Save.CREATE.save(this.getDataModel(), this, SaveAdHoc.class);
+        }
+        else {
+            Save.CREATE.save(this.getDataModel(), this, SaveScheduled.class);
+        }
+    }
+
+    @Override
+    protected void post() {
+        if (this.isAdHoc()) {
+            Save.UPDATE.save(this.getDataModel(), this, SaveAdHoc.class);
+        }
+        else {
+            Save.UPDATE.save(this.getDataModel(), this, SaveScheduled.class);
+        }
     }
 
     @Override
     public void prepareForSaving() {
-        this.nextExecutionSpecs.get().save();
+        if (!this.isAdHoc()) {
+            this.nextExecutionSpecs.get().save();
+        }
         super.prepareForSaving();
     }
 
     @Override
     public void doDelete() {
         super.doDelete();
-        this.nextExecutionSpecs.get().delete();
+        if (!this.isAdHoc()) {
+            this.nextExecutionSpecs.get().delete();
+        }
     }
 
     @Override
-    public boolean isScheduled() {
-        return true;
+    public boolean usesSharedSchedule() {
+        return false;
     }
 
     @Override
@@ -78,7 +123,7 @@ public class ManuallyScheduledComTaskExecutionImpl extends SingleComTaskExecutio
 
     @Override
     public boolean isAdHoc() {
-        return false;
+        return !this.nextExecutionSpecs.isPresent();
     }
 
     @Override
@@ -87,14 +132,23 @@ public class ManuallyScheduledComTaskExecutionImpl extends SingleComTaskExecutio
     }
 
     public void setNextExecutionSpecsFrom(TemporalExpression temporalExpression) {
-        if (!this.nextExecutionSpecs.isPresent()) {
-            this.nextExecutionSpecs.set(this.getSchedulingService().newNextExecutionSpecs(temporalExpression));
+        if (temporalExpression == null) {
+            if (!this.nextExecutionSpecs.isPresent()) {
+                // No change
+            }
+            else {
+                this.nextExecutionSpecs.setNull();
+            }
         }
         else {
-            this.nextExecutionSpecs.get().setTemporalExpression(temporalExpression);
+            if (this.nextExecutionSpecs.isPresent()) {
+                this.nextExecutionSpecs.get().setTemporalExpression(temporalExpression);
+            }
+            else {
+                this.nextExecutionSpecs.set(this.getSchedulingService().newNextExecutionSpecs(temporalExpression));
+            }
         }
     }
-
 
     @Override
     public int getMaxNumberOfTries() {
@@ -102,15 +156,96 @@ public class ManuallyScheduledComTaskExecutionImpl extends SingleComTaskExecutio
     }
 
     @Override
-    // Override from superclass to add @Override annotation
     public ComTask getComTask() {
-        return super.getComTask();
+        // we do an explicit get because ComTask is required and should not be null
+        return comTask.get();
+    }
+
+    private void setComTask(ComTask comTask) {
+        this.comTask.set(comTask);
     }
 
     @Override
-    // Override from superclass only to add @Override
     public ProtocolDialectConfigurationProperties getProtocolDialectConfigurationProperties() {
-        return super.getProtocolDialectConfigurationProperties();
+        return protocolDialectConfigurationProperties.orNull();
+    }
+
+    @Override
+    public List<ComTask> getComTasks() {
+        return Arrays.asList(getComTask());
+    }
+
+    void setProtocolDialectConfigurationProperties(ProtocolDialectConfigurationProperties protocolDialectConfigurationProperties) {
+        this.protocolDialectConfigurationProperties.set(protocolDialectConfigurationProperties);
+    }
+
+    @Override
+    public boolean isConfiguredToCollectRegisterData() {
+        return isConfiguredToCollectDataOfClass(RegistersTask.class);
+    }
+
+    @Override
+    public boolean isConfiguredToCollectLoadProfileData() {
+        return isConfiguredToCollectDataOfClass(LoadProfilesTask.class);
+    }
+
+    @Override
+    public boolean isConfiguredToRunBasicChecks() {
+        return isConfiguredToCollectDataOfClass(BasicCheckTask.class);
+    }
+
+    @Override
+    public boolean isConfiguredToCheckClock() {
+        return isConfiguredToCollectDataOfClass(ClockTask.class);
+    }
+
+    @Override
+    public boolean isConfiguredToCollectEvents() {
+        return isConfiguredToCollectDataOfClass(LogBooksTask.class);
+    }
+
+    @Override
+    public boolean isConfiguredToSendMessages() {
+        return isConfiguredToCollectDataOfClass(MessagesTask.class);
+    }
+
+    @Override
+    public boolean isConfiguredToReadStatusInformation() {
+        return isConfiguredToCollectDataOfClass(StatusInformationTask.class);
+    }
+
+    @Override
+    public boolean isConfiguredToUpdateTopology() {
+        return isConfiguredToCollectDataOfClass(TopologyTask.class);
+    }
+
+    private <T extends ProtocolTask> boolean isConfiguredToCollectDataOfClass (Class<T> protocolTaskClass) {
+        for (ProtocolTask protocolTask : this.getComTask().getProtocolTasks()) {
+            if (protocolTaskClass.isAssignableFrom(protocolTask.getClass())) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    @Override
+    public boolean executesComSchedule(ComSchedule comSchedule) {
+        return false;
+    }
+
+    @Override
+    public boolean executesComTask(ComTask comTask) {
+        return comTask != null && comTask.getId() == this.getComTask().getId();
+    }
+
+    @Override
+    public boolean performsIdenticalTask(ComTaskExecutionImpl comTaskExecution) {
+        return comTaskExecution != null && comTaskExecution.executesComTask(this.getComTask());
+    }
+
+    @Override
+    public List<ProtocolTask> getProtocolTasks() {
+        return Collections.unmodifiableList(this.getComTask().getProtocolTasks());
     }
 
     @Override
@@ -118,16 +253,20 @@ public class ManuallyScheduledComTaskExecutionImpl extends SingleComTaskExecutio
         return new ManuallyScheduledComTaskExecutionUpdaterImpl(this);
     }
 
-    public static class ManuallyScheduledComTaskExecutionBuilderImpl extends AbstractComTaskExecutionBuilder<ManuallyScheduledComTaskExecutionBuilder, ManuallyScheduledComTaskExecution, ManuallyScheduledComTaskExecutionImpl> implements ManuallyScheduledComTaskExecutionBuilder {
+    @Override
+    protected Date calculateNextExecutionTimestamp(Date now) {
+        if (this.isAdHoc()) {
+            return null;
+        }
+        else {
+            return super.calculateNextExecutionTimestamp(now);
+        }
+    }
+
+    public static class ManuallyScheduledComTaskExecutionBuilderImpl extends AbstractComTaskExecutionBuilder<ManuallyScheduledComTaskExecution, ManuallyScheduledComTaskExecutionImpl> implements ComTaskExecutionBuilder<ManuallyScheduledComTaskExecution> {
 
         protected ManuallyScheduledComTaskExecutionBuilderImpl(ManuallyScheduledComTaskExecutionImpl manuallyScheduledComTaskExecution) {
-            super(manuallyScheduledComTaskExecution, ManuallyScheduledComTaskExecutionBuilder.class);
-        }
-
-        @Override
-        public ManuallyScheduledComTaskExecutionBuilder scheduleAccordingTo(TemporalExpression temporalExpression) {
-            this.getComTaskExecution().setNextExecutionSpecsFrom(temporalExpression);
-            return self();
+            super(manuallyScheduledComTaskExecution);
         }
 
     }
@@ -152,6 +291,23 @@ public class ManuallyScheduledComTaskExecutionImpl extends SingleComTaskExecutio
             return self();
         }
 
+        @Override
+        public ManuallyScheduledComTaskExecutionUpdater removeSchedule() {
+            this.getComTaskExecution().setNextExecutionSpecsFrom(null);
+            return self();
+        }
     }
+
+    /**
+     * Uses as a marker interface for javax.validation when
+     * saving an ad-hoc manually scheduled task.
+     */
+    private interface SaveAdHoc {}
+
+    /**
+     * Uses as a marker interface for javax.validation when
+     * saving a manually scheduled task.
+     */
+    private interface SaveScheduled {}
 
 }
